+++ zench/src/report/mod.rs
@@ -3,6 +3,8 @@
 mod colors;
 mod display;
 mod report;
+#[cfg(not(feature = "display_vertical"))]
+mod table;
 
 pub use report::Report;